Add compatibility for S3-compliant APIs
This commit adds support for configuring a S3-compatible backend, allowing
this sample app to be used on other object storage platforms. To do so,
I added three additional (and optional) configuration parameters:

- endpoint: the S3-compatible API endpoint (e.g. https://my-s3-backend.com:1234).
If not set, it will AWS S3.

- pathStyleAccess: if set to true, it will use path style URLs. If set to
false, it will use virtual host style URLs (bucket in subdomain).
Defaults to false as it's the default value for AWS.

- baseUrl: used to generate the public link of the objects uploaded (i.e.
{baseUrl}/{bucket}/{object}). If not set, it will generate the URL using
the getUrl method of the client SDK.

As the project hasn't been updated in 3 years, I also updated the
dependencies and the code to comply with the new versions. The region
parameter is now mandatory as it's needed by the new version of the AWS
Java Client.

I also updated the UI dependencies (Bootstrap, Jquery) and some HTML +
CSS to make it look a bit nicer.